% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{lc_palettes}
\alias{lc_palettes}
\alias{lc_terrain}
\alias{lc_topo}
\alias{lc_heat}
\title{Predefined palettes and palette generators}
\usage{
lc_terrain(n)

lc_topo(n)

lc_heat(n)
}
\arguments{
\item{n}{if not provided, return the vector of official Lego colors defining
the palette. If provided, return an interpolated palette, which will contain
non-Lego colors only if \code{n} is greater than the palette definition.}
}
\value{
character vector of hex colors
}
\description{
These functions are analogs to familiar R color palettes but based on Lego
colors.
}
\examples{
lc_terrain()
lc_terrain(4)

lc_topo()
lc_heat()
}
