% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legocolors.R
\docType{package}
\name{legocolors}
\alias{legocolors-package}
\alias{legocolors}
\title{legocolors: Official Lego Color Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a dataset containing several color naming conventions established by multiple sources, along with associated color metadata. The package also provides related helper functions for mapping among the different Lego color naming conventions and between Lego colors, hex colors, and 'R' color names, making it easy to convert any color palette to one based on existing Lego colors while keeping as close to the original color palette as possible. The functions use nearest color matching based on Euclidean distance in RGB space. Naming conventions for color mapping include those from 'BrickLink' (\url{https://www.bricklink.com}), 'The Lego Group' (\url{https://www.lego.com}), 'LDraw' (\url{https://www.ldraw.org/}), and 'Peeron' (\url{http://www.peeron.com/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonawicz/legocolors}
  \item Report bugs at \url{https://github.com/leonawicz/legocolors/issues}
}

}
\author{
\strong{Maintainer}: Matthew Leonawicz \email{rpkgs@pm.me}

}
