% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legoCols}
\alias{legoCols}
\title{Lego color data.}
\format{
A data frame with 12 columns and 167 rows.
}
\usage{
legoCols
}
\description{
A data frame containing Lego color data.
}
\keyword{datasets}
