% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legoPals}
\alias{legoPals}
\title{Lego color palettes.}
\format{
A named list of nine color palettes.
}
\usage{
legoPals
}
\description{
A named list of nine Lego color palettes based on \code{legoCols$material}.
Because these palettes represent a meaningful segmentation of Lego material types,
they are not intended to represent useful color palettes for mapping data values to colors.
}
\keyword{datasets}
