% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legocolors.R
\docType{data}
\name{legocolors}
\alias{legocolors}
\title{Lego color data.}
\format{A data frame with 13 columns and 165 rows.}
\usage{
legocolors
}
\description{
A data frame containing Lego color data.
}
\keyword{datasets}
