% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legocolors.R
\docType{data}
\name{legopals}
\alias{legopals}
\title{Lego color data.}
\format{A named list of nine color palettes.}
\usage{
legopals
}
\description{
A named list of nine Lego color palettes based on `legocolors$material`.
}
\keyword{datasets}
